section 10

+++ Application/Activities/Delete.cs
@@ -1,3 +1,4 @@
+using Application.Core;
 using MediatR;
 using Persistence;
 
@@ -5,11 +6,12 @@ namespace Application.Activities
 {
     public class Delete
     {
-        public class Command : IRequest{
+        public class Command : IRequest<Result<Unit>>
+        {
             public Guid Id { get; set; }
         }
 
-        public class Handler : IRequestHandler<Command>
+        public class Handler : IRequestHandler<Command, Result<Unit>>
         {
             private readonly DataContext _context;
 
@@ -17,17 +19,22 @@ namespace Application.Activities
             {
                 _context = context;
             }
-            public  async Task<Unit> Handle(Command request, CancellationToken cancellationToken)
+            public  async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
             {
                 
                 // _context.Activities.Remove(request.Activity);
                 // the code above don't work directly because the api don't have any body
                 var activity= await _context.Activities.FindAsync(request.Id);
+
+                if(activity == null) return null;
+
                 _context.Activities.Remove(activity);
 
-                await _context.SaveChangesAsync();
+                var result = await _context.SaveChangesAsync() > 0;
 
-                return Unit.Value;
+                if(!result) return Result<Unit>.Failure("Failed to delete the activity");
+
+                return Result<Unit>.Success(Unit.Value);
             }
         }
     }
